criando a funcao changeColors dentro de services e importando dentro da Questions
